requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.2
Pillow>=8.1.0
matplotlib>=3.3.2
tqdm>=4.59.0
PyYAML>=5.4.1
scikit-image>=0.18.1
opencv-python>=4.5.1
tensorboard>=2.4.0
timm>=0.6.0 